Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (1 comment)
> > 
> > Motivation for this is to get this package build in CentOS / RHEL 8 in CBS,
> > we have not had it yet https://cbs.centos.org/koji/packageinfo?packageID=6817
> > since some static libs are not published in RHEL 8.
> > 
> > @Javier how does DLRN gets missing static libs?
> 
> We've been adding them manually to the repo. @Alfredo can you share some more details?

I rebuild it with mock using the distgit and sources from git.centos.org and push the openssl-static in the build-deps repo:

https://trunk.rdoproject.org/centos8-master/build-deps/latest/other/

This is fine for RDO Trunk, but not for cloudsig. For cloudsig we are simply skipping this build. We requested CentOS team to include this package in the devel repo with no success. Acoording to latest news from them, there will be some announcement about a new process fro missing subpackages soon. We'll see if it helps.

Patch-set: 1
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>