Use %gobuild from go-rpm-macros for golang httpd test binary
Using the platform-defined %gobuild lets us use all the compile and
linking flags recommended by the platform rather than having to hardcode
them in our specfile, with the potential of becoming stale.

This came up as the executable being compiled with CGO_ENABLED was
flagged as potentially not FIPS compliant, because it couldn't be linked
against kown-compliant system-provided crypto libraries.

Change-Id: I9169646b279b70c8c2d9a02df3a17c5d785a8157